[Fix][Antonio] User logging moved to user navbar
